remove calls to java stream() (#37539)
## What
<!--
* Describe what the change is solving. Link all GitHub issues related to this change.
-->

## How
<!--
* Describe how code changes achieve the solution.
-->

## Review guide
<!--
1. `x.py`
2. `y.py`
-->

## User Impact
<!--
* What is the end result perceived by the user?
* If there are negative side effects, please list them. 
-->

## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [ ] YES 💚
- [ ] NO ❌

+++ airbyte-cdk/java/airbyte-cdk/db-sources/src/testFixtures/kotlin/io/airbyte/cdk/integrations/standardtest/source/TestDataHolder.kt
@@ -17,7 +17,7 @@ internal constructor(
     private val fullSourceDataType: String?
 ) {
     var nameSpace: String? = null
-    private var testNumber: Long = 0
+    private var testNumber: Int = 0
     private var idColumnName: String? = null
     private var testColumnName: String? = null
 
@@ -154,7 +154,7 @@ internal constructor(
         }
     }
 
-    fun setTestNumber(testNumber: Long) {
+    fun setTestNumber(testNumber: Int) {
         this.testNumber = testNumber
     }
